Significantly improve QR code detection accuracy
Use a second method (library) for scanning the QR code. Each of the two might work better under certain conditions, therefore decreasing the failrure rate.

The initial library used needed very specific lighting conditions and angles for it to work. This second library takes over when the first one fails. It might be a bit less accurate in terms of the dimensions of the QR code, but it works a lot more often.